+++ wandb/integration/keras/callbacks/model_checkpoint.py
@@ -29,7 +29,7 @@ class WandbModelCheckpoint(callbacks.ModelCheckpoint):
     This callback is to be used in conjunction with training using `model.fit()` to save
     a model or weights (in a checkpoint file) at some interval. The model checkpoints
     will be logged as W&B Artifacts. You can learn more here:
-    https://docs.wandb.ai/guides/artifacts
+    https://docs.wandb.ai/models/artifacts
 
     This callback provides the following features:
         - Save the model that has achieved "best performance" based on "monitor".