+++ wandb/sdk/wandb_sweep.py
@@ -43,7 +43,7 @@ def sweep(
     Make note the unique identifier, `sweep_id`, that is returned.
     At a later step provide the `sweep_id` to a sweep agent.
 
-    See [Sweep configuration structure](https://docs.wandb.ai/guides/sweeps/define-sweep-configuration)
+    See [Sweep configuration structure](https://docs.wandb.ai/models/sweeps/define-sweep-configuration)
     for information on how to define your sweep.
 
     Args: